+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to
 - The canonical format (`-Tcanon`) produces output that is stable across
   multiple invocations. #2193
 
+- Invisible edges no longer force visible parallel edges onto separate spline
+  lanes. #1093
+
 ## [16.1.0] – 2026-09-03
 
 ### Added